requirements.txt
streamlit
openai
qdrant-client
Pillow
python-dotenv